+++ core/src/animation/progress.rs
@@ -1,4 +1,4 @@
-#[derive(PartialEq, Copy, Clone)]
+#[derive(PartialEq, Copy, Clone, Debug)]
 pub enum AnimateProgress {
   Dismissed,
   Between(f32),