+++ widgets/src/lib.rs
@@ -11,6 +11,7 @@ pub mod layout;
 pub mod link;
 pub mod lists;
 pub mod path;
+pub mod progress;
 pub mod radio;
 pub mod scrollbar;
 pub mod tabs;
@@ -20,7 +21,7 @@ pub mod transform_box;
 pub mod prelude {
   pub use super::{
     avatar::*, buttons::*, checkbox::*, common_widget::*, divider::*, grid_view::*, icon::*,
-    input::*, label::*, layout::*, link::*, lists::*, path::*, radio::*, scrollbar::*, tabs::*,
-    text::*, text_field::*, transform_box::*,
+    input::*, label::*, layout::*, link::*, lists::*, path::*, progress::*, radio::*, scrollbar::*,
+    tabs::*, text::*, text_field::*, transform_box::*,
   };
 }